# Redo minor cleanups in System.Runtime.Numerics