--- a/src/libraries/System.Runtime.Numerics/src/System/Numerics/BigInteger.cs
+++ b/src/libraries/System.Runtime.Numerics/src/System/Numerics/BigInteger.cs
@@ -1085,8 +1085,8 @@ namespace System.Numerics
                 return _sign;
 
             HashCode hash = default;
-            hash.Add(_sign);
             hash.AddBytes(MemoryMarshal.AsBytes(_bits.AsSpan()));
+            hash.Add(_sign);
             return hash.ToHashCode();
         }