Redo a part of https://github.com/dotnet/runtime/pull/53984
I don't expect it will cause regression this time. Should I run the benchmarks?