@stephentoub 2022-06-24
> Should I run the benchmarks?

Yes, please.